Add proto packet listener API. Add mime-type database for later use by analyzers. Use size_t for HTTP Content-Length. Add analyzer payload API and basic support in analyzer_http.

+++ src/modules/proto/proto_http.h
@@ -42,7 +42,7 @@
 
 
 struct http_info {
-	unsigned int content_len, content_pos;
+	size_t content_len, content_pos;
 	unsigned int flags;
 };
